Add Estimator
Now, EPBO is complete:
1. add estimator in pbort.c
2. add XXX_est in each component
3. test ESTIMATOR in manageUnit.c

+++ ReFrESH_SourceCodes/Bubble1_to_Bubble4/Bubble1_to_Bubble4/src/actuator.h
@@ -33,7 +33,7 @@ typedef struct{
 	 * and 2) non-functional evaluated value. These values are compared with the
 	 * status of the node. */
 	uint8_t				funcPerfValue;		/* functional performance value of each component */
-	uint8_t				power;
+	int					power;
 	int					commuRSSI;
 
 	/* TODO:
